+++ src/main/java/com/example/tripplanner/service/UserService.java
@@ -1,10 +1,13 @@
 package com.example.tripplanner.service;
 
+import com.example.tripplanner.entities.Destination;
 import com.example.tripplanner.entities.User;
 import com.example.tripplanner.repositories.UserRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
-import java.util.Optional;
+import java.util.List;
 
 @Service
 public class UserService {
@@ -19,19 +22,32 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public Optional<User> getUserById(Long id){
-        return userRepository.findById(id);
+    public User getUserById(Long id){
+        return userRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
     }
 
-    public Optional<User> getUserByUsername(String username){
-        return userRepository.findByUsername(username);
-    }
-
-    public void deleteUser(User user){
-        userRepository.delete(user);
+    public User getUserByUsername(String username){
+        return userRepository.findByUsername(username).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
     }
 
     public void deleteUserById(Long id){
-        userRepository.deleteById(id);
+        if(userRepository.existsById(id))
+            userRepository.deleteById(id);
+        else
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+    }
+
+    public List<User> findAll() {
+        return userRepository.findAll();
+    }
+
+    public User updateUser(Long id, User user) {
+        if (userRepository.existsById(id)) {
+            user.setId(id);
+
+            return userRepository.save(user);
+        } else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
     }
 }
